use niv and naersk for Nix packaging
[naersk](https://github.com/nmattia/naersk) provides a Nix expression to
build Cargo crates without using a fixed-output derivation. This brings
many advantages compared to the rustPlatform expression from nixpkgs:
- It is not necessary to keep the cargoSha256 up-to-date
- Metadata is read from Cargo.toml and does not have to be maintained
  twice
- All dependencies are built in a seperate derivation and do not have to
  be rebuilt with every change in shelfie

I chose to use niv to have a fixed reference to naersk and nixpkgs,
so that users do not have to worry about incompatible changes in newer
versions if they do not want to.